+++ front/cypress/e2e/proposals_permissions/post_participation.ts
@@ -1,6 +1,6 @@
-import { randomString } from '../../../support/commands';
-import { signUpEmailConformation, enterUserInfo } from '../../../support/auth';
-import { updatePermission } from '../../../support/permitted_by_utils';
+import { randomString } from '../../support/commands';
+import { signUpEmailConformation, enterUserInfo } from '../../support/auth';
+import { updatePermission } from '../../support/permitted_by_utils';
 
 describe('Post Participation Signup: proposals', () => {
   let projectId: string;
