+++ front/cypress/e2e/proposals_permissions/project_proposal_page_action.cy.ts
@@ -1,4 +1,4 @@
-import { randomString, randomEmail } from '../support/commands';
+import { randomString, randomEmail } from '../../support/commands';
 import moment = require('moment');
 
 describe('Proposal show page actions', () => {
